python_backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
requests
python-dotenv
google-generativeai>=0.5.0
